% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_arm_functions.R
\name{add_chromosome_arms}
\alias{add_chromosome_arms}
\title{Add chromosome arm information to GRCh38 region coordinates}
\usage{
add_chromosome_arms(
  df,
  chrom_col = chromosome_char,
  start_col = start,
  end_col = end
)
}
\arguments{
\item{df}{A dataframe containing region information.}
}
\value{
The input dataframe with an additional "chromosome_arm" column for
autosome data only. Sex chromosome coordinates are removed.
}
\description{
Chromosome arm coordinates are taken from this github repo:
https://github.com/Neurosurgery-Brain-Tumor-Center-DiazLab/CONICS/blob/master/chromosome_arm_positions_grch38.txt
}
