% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_functions.R
\name{add_identifiers}
\alias{add_identifiers}
\title{Add identifiers in a filename onto a dataframe}
\usage{
add_identifiers(file, tbl)
}
\arguments{
\item{file}{The filename with identifiers}

\item{tbl}{The table to add identifiers to}
}
\value{
A dataframe of the initial table joined to the patient identifiers
}
\description{
Add identifiers in a filename onto a dataframe
}
\examples{
filename <- "Annotated_WS123456_12345678a_JoeShaw.xlsx"

data <- data.frame(gene = c("ERBB2"), dosage = c(2))

data_with_ids <- add_identifiers(file = filename, tbl = data)
}
