% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_metric_functions.R
\name{calc_sensitivity}
\alias{calc_sensitivity}
\title{Calculate the sensitivity for a test in comparison to another test}
\usage{
calc_sensitivity(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
The sensitivity (true positive rate) of the test as a percentage
}
\description{
Calculate the sensitivity for a test in comparison to another test
}
\examples{
calc_sensitivity(tp = 40, tn = 44, fn = 3, fp = 0)
}
