% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_metric_functions.R
\name{calc_specificity}
\alias{calc_specificity}
\title{Calculate the specifity for a test in comparison to another test}
\usage{
calc_specificity(tp, tn, fp, fn)
}
\arguments{
\item{tp}{True positives}

\item{tn}{True negatives}

\item{fp}{False positives}

\item{fn}{False negatives}
}
\value{
The specificity (true negative rate) as a percentage
}
\description{
Calculate the specifity for a test in comparison to another test
}
\examples{
calc_specificity(tp = 40, tn = 44, fn = 3, fp = 0)
}
