% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_oligos_in_range.R
\name{count_oligos_in_range}
\alias{count_oligos_in_range}
\title{Count QIAseq oligos within a genomic coordinate range}
\usage{
count_oligos_in_range(chrom, coord1, coord2, df)
}
\arguments{
\item{chrom}{The chromosome of the range of interest}

\item{coord1}{The first coordinate of the range of interest}

\item{coord2}{The second coordinate of the range of interest}

\item{df}{The dataframe containing the oligonucleotide information.
This should be in the format of one oligo (primer or target) per row with one column each for the
start and end coordinates.}
}
\value{
Returns the number of oligos within the specified region
}
\description{
"Oligo" refers to an oligonucleotide, which may be a QIAseq primer or target. A QIAseq target is
a region of a BED file targeted by multiple primers.
}
\details{
This function can be used with dplyr::rowwise() for a dataframe input.
}
