% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erbb2_functions.R
\name{csv_timestamp}
\alias{csv_timestamp}
\title{Save a table as a comma-separated file with a timestamp}
\usage{
csv_timestamp(table, folder)
}
\arguments{
\item{table}{The table to save}

\item{folder}{The filepath of the folder to save the file into, without a
backlash at the end}
}
\value{
Saves the table in the desired folder with a timestamp
}
\description{
Save a table as a comma-separated file with a timestamp
}
\note{
This function is used for exporting tables from RStudio for inclusion
in validation documentation.
}
