% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_arm_functions.R
\name{define_chromosome_levels}
\alias{define_chromosome_levels}
\title{Define factor levels for human chromosomes}
\usage{
define_chromosome_levels()
}
\value{
A vector of human autosomes and sex chromosomes in numeric order
and then X and Y.
}
\description{
This function can be used to define the factor levels of a chromosome
variable so that the observations will sort in chromosome order.
}
