% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom_functions.R
\name{define_uom}
\alias{define_uom}
\title{Define the uncertainty of measurement at a given confidence level}
\usage{
define_uom(pooled_sd, num_sd)
}
\arguments{
\item{pooled_sd}{The pooled standard deviation}

\item{num_sd}{The number of standard deviations from the mean, which defines the confidence level}
}
\value{
The uncertainty of measurement value.
}
\description{
When a normal (Gaussian) distribution is assumed, a defined number of results will fall within
different distances from the mean, which can be described by \href{https://en.wikipedia.org/wiki/Normal_distribution#/media/File:Standard_deviation_diagram_micro.svg}{standard deviations}.
}
\details{
\itemize{
\item ± 1 standard deviation: 68.2\% of results
\item ± 2 standard deviations: 95.4\% of results
\item ± 3 standard deviations: 99.6\% of results
}
}
\examples{
define_uom(0.0781, 2)
}
