% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erbb2_functions.R
\name{extract_cnv_calls}
\alias{extract_cnv_calls}
\title{Extract CNV calls from DNA Database genotype comments}
\usage{
extract_cnv_calls(df, input_gene)
}
\arguments{
\item{df}{A dataframe including a "genotype" column of QIAseq results from the
DNA Database.}

\item{input_gene}{The gene of interest}
}
\value{
A dataframe including the input dataframe with additional columns
generated from the "genotype" column for the gene CNV result and dosage quotient.
}
\description{
Extract CNV calls from DNA Database genotype comments
}
\note{
Genotypes are entered onto the DNA Database as a text string with
a consistent structure. This function parses results for a
}
