% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{extract_cnv_tbls}
\alias{extract_cnv_tbls}
\title{Extract all information from the CNV sheet of a PanSolid CLC Excel output}
\usage{
extract_cnv_tbls(
  filepath,
  sheet_regex = "CNVs_",
  stdev_string = "StDev Signal-adjusted Log2 Ratios",
  percent_138x_string = "\% Whole Panel Covered at 138X",
  pred_ncc_string = "Predicted NCC (\%)",
  num_amp_genes = 13,
  amp_string = "Amplification genes",
  num_del_genes = 34,
  del_string = "Deletion genes",
  del_col_length = 15,
  sig_cnv_string = "Significant CNV results"
)
}
\arguments{
\item{filepath}{Full file path to an Excel file}

\item{sheet_regex}{Regular expression for sheet name matching}
}
\value{
A named list of tables
}
\description{
This function acts as a wrapper of the \verb{find_*} functions to extract
information from the 5 tables on the CLC pipeline Excel output CNV
sheet, adds sample identifiers and stores them within a list.
}
\details{
The defaults are set for the version of the CNV tab used in the live
PanSolid service following addition of deletions and LOH.
}
