% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pansolid_cnv_coordinates.R
\name{extract_pansolid_cnv_coordinates}
\alias{extract_pansolid_cnv_coordinates}
\title{Extract coordinates from the PanSolid CNV coordinate string}
\usage{
extract_pansolid_cnv_coordinates(df, cnv_coord_col)
}
\arguments{
\item{df}{A dataframe containing a column of CNV coordinates to parse.
This function was designed mainly to be used with the "Positive/Significant
CNV results" table in the PanSolid Excel output.}

\item{cnv_coord_col}{The column in the dataframe containing the CNV
coordinates separated by two full stops (Example: "15889325..15946102")}
}
\value{
The input dataframe with the CNV start and end coordinates
as two new columns.
}
\description{
Extract coordinates from the PanSolid CNV coordinate string
}
\examples{

df <- data.frame("cnv_coord" = c("55174776..55174793"))

cnvs_with_coordinates <- extract_pansolid_cnv_coordinates(df, 
cnv_coord_col = cnv_coord)
}
