% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_arm_functions.R
\name{factorise_chromosome}
\alias{factorise_chromosome}
\title{Factorise the chromosome column in a dataframe}
\usage{
factorise_chromosome(df, col)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The column of the dataframe with chromosome information of
type character.}
}
\value{
The input dataframe with a new column called "chromosome_fct"
containing the factorised version of the chromosome.
}
\description{
This function is designed to make sorting by chromosome easier.
If chromosome information is formatted as a character then sorting does
not work with autosomes (chromosome "10" is sorted after chromosome "1", not
after chromosome "9").
If chromosome information is formatted as a numeric then sex chromosomes
cannot be correctly formatted.
Formatting as a factor allows easier by specifying the order of autosomes
first, then sex chromosomes.
}
