% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_functions.R
\name{filename_to_df}
\alias{filename_to_df}
\title{Convert PanSolid results Excel filename identifiers to a dataframe}
\usage{
filename_to_df(file)
}
\arguments{
\item{file}{The filename or full filepath}
}
\value{
A dataframe of patient filename identifiers
}
\description{
Convert PanSolid results Excel filename identifiers to a dataframe
}
\examples{
filename <- "Annotated_WS123456_12345678a_JoeShaw.xlsx"

id_df <- filename_to_df(filename)
}
