% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{find_amp_genes}
\alias{find_amp_genes}
\title{Find fold change information for amplification genes with the CNV sheet}
\usage{
find_amp_genes(input_sheet, num_genes = 9, amp_string = "Amplification genes")
}
\arguments{
\item{input_sheet}{A tibble containing the amplified gene information.
This is intended to be the output from \code{read_cnv_sheet}.}

\item{num_genes}{The number of genes with amplification results returned
by the CLC pipeline. This is currently 9 but may change in future.}

\item{amp_string}{The string which identifies the position of the
amplification genes table. This defaults to the current header, but
can be changed if necessary.}
}
\value{
A tibble of the amplification genes results table
}
\description{
Find fold change information for amplification genes with the CNV sheet
}
