% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{find_del_genes}
\alias{find_del_genes}
\title{For fold change information for deletion genes within the CNV sheet}
\usage{
find_del_genes(
  input_sheet,
  num_genes = 37,
  del_string = "Deletion genes",
  col_length = 15
)
}
\arguments{
\item{input_sheet}{A tibble containing the deleted gene information.
This is intended to be the output from \code{read_cnv_sheet}.}

\item{num_genes}{The number of genes within the deletion genes table.}

\item{del_string}{The string which identifies the position of the
deletion genes table. This defaults to the current header, but
can be changed if necessary.}

\item{col_length}{The length in rows of the columns that genes are
presented in. Defaults to 15 and then blank space is removed
automatically.}
}
\value{
A tibble of the deletion genes results table
}
\description{
There are currently 37 genes identified for deletion analysis within the
CLC pipeline. These are presented in 3 tables. \code{find_del_genes} identifies
each table within the CNV sheet and collates them together.
}
