% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{find_match}
\alias{find_match}
\title{Find matches for strings within an Excel sheet}
\usage{
find_match(input_sheet, input_col, match_string)
}
\arguments{
\item{input_sheet}{The Excel sheet to search}

\item{input_col}{The column to search}

\item{match_string}{The string to search for}
}
\value{
The numeric position of the string match within the column
}
\description{
Find matches for strings within an Excel sheet
}
