% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{find_percent_138x}
\alias{find_percent_138x}
\title{Find percentage coverage at 138X within the CNV sheet}
\usage{
find_percent_138x(
  input_sheet,
  percent_138x_string = "\% Whole Panel Covered at 138X"
)
}
\arguments{
\item{input_sheet}{A tibble containing the percent 138X information.
This is intended to be the output from \code{read_cnv_sheet}.}

\item{percent_138x_string}{The string which identifies the position of the
percent coverage at 138x table. This defaults to the current header, but
can be changed if necessary.}
}
\value{
A tibble with the percent 138X information.
}
\description{
Find percentage coverage at 138X within the CNV sheet
}
