% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{find_pred_ncc}
\alias{find_pred_ncc}
\title{Find predicted neoplastic cell content (NCC) within the CNV sheet}
\usage{
find_pred_ncc(input_sheet, pred_ncc_string = "Predicted NCC (\%)")
}
\arguments{
\item{input_sheet}{A tibble containing the percent NCC information.
This is intended to be the output from \code{read_cnv_sheet}.}

\item{pred_ncc_string}{The string which identifies the position of the
predicted NCC value. This defaults to the current header, but
can be changed if necessary.}
}
\value{
A tibble with the predicted NCC information.
}
\description{
Find predicted neoplastic cell content (NCC) within the CNV sheet
}
