% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{find_sig_cnvs}
\alias{find_sig_cnvs}
\title{Find information for significant CNV results within the CNV sheet}
\usage{
find_sig_cnvs(
  input_sheet,
  sig_cnv_string = "Significant CNV results",
  amp_gene_string = "Amplification genes"
)
}
\arguments{
\item{input_sheet}{A tibble containing the significant CNV information.
This is intended to be the output from \code{read_cnv_sheet}.}

\item{sig_cnv_string}{The string which identifies the position of the
significant CNVs table. This defaults to the current header, but
can be changed if necessary.}

\item{amp_gene_string}{The string which identifiers the position
of the amplification genes table.}
}
\value{
A tibble of the significant CNV results table
}
\description{
Find information for significant CNV results within the CNV sheet
}
