% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{find_stdev_ratios}
\alias{find_stdev_ratios}
\title{Find standard deviation of signal-adjusted log2 ratios within the CNV sheet}
\usage{
find_stdev_ratios(
  input_sheet,
  stdev_string = "StDev Signal-adjusted Log2 Ratios"
)
}
\arguments{
\item{input_sheet}{A tibble containing the stdev information. This is
intended to be the output from \code{read_cnv_sheet}.}

\item{stdev_string}{The string which identifies the position of the
standard deviation information. This defaults to the current header, but
can be changed if necessary.}
}
\value{
A tibble with the standard deviation of signal-adjusted
log2 ratios
}
\description{
Find standard deviation of signal-adjusted log2 ratios within the CNV sheet
}
