% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erbb2_functions.R
\name{format_chromosome}
\alias{format_chromosome}
\title{Format a chromosome string}
\usage{
format_chromosome(df, input_col)
}
\arguments{
\item{df}{Input dataframe}

\item{input_col}{Column containing chromosome information}
}
\value{
The input dataframe with an additional column which presents chromosome
information as a single character string. Autosomes are presented as characters
without decimal places (i.e. "1" instead of "1.0")
}
\description{
Format a chromosome string
}
