% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_arm_functions.R
\name{format_chromosome_decimals}
\alias{format_chromosome_decimals}
\title{Format chromosome character strings when formatted as decimal numbers}
\usage{
format_chromosome_decimals(df, col = chromosome)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The column of the dataframe containing
chromosome information formatted as a decimal number (i.e. "1.0").}
}
\value{
The dataframe with an additional "chromosome_char" column with
the decimal places removed.
}
\description{
PanSolid Excel outputs contain columns with chromosome information.
When these columns are read as text, the autosomes are converted into
strings with decimal places - example: "1" becomes "1.0".
\code{format_chromosome_decimals} removes the decimal place to make these
strings easier to work with
}
