% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{get_amp_sheetname}
\alias{get_amp_sheetname}
\title{Get the sheet name of the amplifications tab for a PanSolid results Excel}
\usage{
get_amp_sheetname(filepath)
}
\arguments{
\item{filepath}{The full filename of the Excel}
}
\value{
A character string of the amplifications tab name
}
\description{
Get the sheet name of the amplifications tab for a PanSolid results Excel
}
\note{
Prior to go-live, the results Excels defaulted to "Amplifications" as the
name of the tab containing amplification information. After go live this changed
to include the sample DNA number (see example).
}
