% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{get_annotated_filepaths}
\alias{get_annotated_filepaths}
\title{Get the filepaths of PanSolid results Excels from the S drive}
\usage{
get_annotated_filepaths(
  repository_path = "S:/central shared/Genetics/Repository/WorksheetAnalysedData/",
  worksheet,
  full_names = TRUE
)
}
\arguments{
\item{repository_path}{The filepath for the worksheet repository - defaults to
the standard S drive location.}

\item{worksheet}{The PanSolid worksheet}

\item{full_names}{TRUE or FALSE}
}
\value{
A list of filepaths
}
\description{
Get the filepaths of PanSolid results Excels from the S drive
}
\note{
PanSolid results Excels are automatically saved onto the S drive in a defined
folder structure
}
