% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{get_cnv_plot_x_breaks}
\alias{get_cnv_plot_x_breaks}
\title{Get consistent X axis breaks for creating CNV plots}
\usage{
get_cnv_plot_x_breaks(interval, plot_xmin, plot_xmax)
}
\arguments{
\item{interval}{The desired interval between breaks}

\item{plot_xmin}{The desired x axis minimum value}

\item{plot_xmax}{The desired x axis maximum value}
}
\value{
A list of breaks
}
\description{
Get consistent X axis breaks for creating CNV plots
}
