% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{get_cnv_plot_xmax}
\alias{get_cnv_plot_xmax}
\title{Get the X axis maximum value for a CNV plot}
\usage{
get_cnv_plot_xmax(df, buffer)
}
\arguments{
\item{df}{A dataframe containing CNVs}

\item{buffer}{The desired buffer to add onto the width of the X axis.}
}
\value{
The value for the X axis maximum
}
\description{
Get the X axis maximum value for a CNV plot
}
