% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{get_cnv_plot_xmin}
\alias{get_cnv_plot_xmin}
\title{Get the X axis minimum value for a CNV plot}
\usage{
get_cnv_plot_xmin(df, buffer)
}
\arguments{
\item{df}{A dataframe containing CNVs}

\item{buffer}{The desired buffer to add onto the width of the X axis. This can be
adjusted to preference based on how close or far from the X axis minimum the CNV
calls should be.}
}
\value{
The value for the X axis minimum.
}
\description{
Get the X axis minimum value for a CNV plot
}
