% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{get_data_for_cnv_plot}
\alias{get_data_for_cnv_plot}
\title{Get data for making a copy number variant plot}
\usage{
get_data_for_cnv_plot(df, gene)
}
\arguments{
\item{df}{The dataframe containing CNV information to be plotted. This should be based
on the "Positive CNV results" table from the PanSolid Excel files.}

\item{gene}{The gene of interest}
}
\value{
Returns the dataframe filtered to include CNVs in the gene of interest
}
\description{
Get data for making a copy number variant plot
}
