% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_metric_functions.R
\name{get_epir_metrics}
\alias{get_epir_metrics}
\title{Get test metrics using epiR}
\usage{
get_epir_metrics(outcome_vector)
}
\arguments{
\item{outcome_vector}{Numeric vector in the order: true positives, false positives,
false negatives, true negatives}
}
\value{
A dataframe with test metrics formatted for publication.
}
\description{
Get test metrics using epiR
}
