% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{get_full_amp_sheet}
\alias{get_full_amp_sheet}
\title{Read the full sheet of the "Amplifications" tab of a PanSolid results Excel}
\usage{
get_full_amp_sheet(file, sheet = "Amplifications")
}
\arguments{
\item{file}{The full filepath of the PanSolid results file}

\item{sheet}{The name of the sheet to read}
}
\value{
All the information in the "Amplifications" tab as a data-frame.
}
\description{
Read the full sheet of the "Amplifications" tab of a PanSolid results Excel
}
