% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{get_gene_chromosome}
\alias{get_gene_chromosome}
\title{Get the chromosome for a gene}
\usage{
get_gene_chromosome(gene)
}
\arguments{
\item{gene}{The gene of interest}
}
\value{
The chromosome which the gene is on, formatted as a string.
}
\description{
Get the chromosome for a gene
}
