% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{get_sheetname}
\alias{get_sheetname}
\title{Get sheet names from an Excel file using regular expressions}
\usage{
get_sheetname(filepath, sheet_regex = "CNVs_")
}
\arguments{
\item{filepath}{Full file path to an Excel file}

\item{sheet_regex}{Regular expression for sheet name matching}
}
\value{
The name of the sheet as a string
}
\description{
Excel files exported from the CLC pipeline have each sheet named with
the 8 digit DNA laboratory number as a suffix.
Example: "Amplifications_12345678"
\code{get_sheetname} can be used to find the specific sheet name for a file.
}
