% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{get_worksheet_filepaths}
\alias{get_worksheet_filepaths}
\title{Get the filepaths of files within worksheet folders}
\usage{
get_worksheet_filepaths(
  worksheet,
  repository_path = "S:/central shared/Genetics/Repository/WorksheetAnalysedData/",
  file_regex = "^Annotated(_|_v2.+_)WS\\\\d{6}_.+.xlsx",
  full_names = TRUE,
  recursive = TRUE
)
}
\arguments{
\item{worksheet}{The PanSolid worksheet in the format WS123456}

\item{repository_path}{The filepath for the worksheet repository - defaults to
the standard S drive location.}

\item{file_regex}{The regular expression for the filetype to find, which
defaults to the annotated output of the PanSolid pipeline}

\item{full_names}{TRUE or FALSE supplied to the full.names argument of
\code{list.files}.}

\item{recursive}{TRUE or FALSE supplied to the recursive argument of
\code{list.files}.}
}
\value{
A list of filepaths
}
\description{
This function is a more flexible version of \code{get_annotated_filepaths}, as
it has the option to specify a different filetype regular expression, and
also options for recursive and full.names arguments in \code{list.files}.
}
