% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom_functions.R
\name{group_sd}
\alias{group_sd}
\title{Calculate standard deviations with grouped variables}
\usage{
group_sd(df, group_variable1, group_variable2, measurement_variable)
}
\arguments{
\item{df}{A dataframe containing measurement data}

\item{group_variable1}{First variable to group by}

\item{group_variable2}{Second variable to group by}

\item{measurement_variable}{The variable to calculate the standard deviation for}
}
\value{
A grouped data frame showing standard deviation
}
\description{
This function performs calculations described in
\href{https://nvlpubs.nist.gov/nistpubs/ir/2019/NIST.IR.6969-2019.pdf}{NISTIR 6969}
Selected Laboratory and Measurement Practices and Procedures to Support
Basic Mass Calibrations (2019 Ed)]
(page 211, section 8.4).
}
