% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dna_submission_sheets.R
\name{join_pansolid_submission_sheets}
\alias{join_pansolid_submission_sheets}
\title{Load and join PanSolid DNA submission sheets}
\usage{
join_pansolid_submission_sheets()
}
\value{
A tidy dataframe for all PanSolid submissions from 2022 to 2024, based on
files saved in the local drive.
}
\description{
Load and join PanSolid DNA submission sheets
}
\note{
Excel spreadsheets for coordinating testing on the PanSolid QIAseq
enrichment are manually curated by the tech team and stored either on the S
drive or on the laboratory Sharepoint. This function uses versions of
these Excel spreadsheets copied onto a local drive.

The sample_id field has some surprises. Example: 23024772 has a degree sign (°)
entered after it which is invisible in Excel and R.
}
