% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_table_functions.R
\name{load_pansolid_gene_table}
\alias{load_pansolid_gene_table}
\title{Load a PanSolid copy number variant gene table}
\usage{
load_pansolid_gene_table(cnv_type)
}
\arguments{
\item{cnv_type}{The type of copy number variant, which must be "Deletions" or "Amplifications"}
}
\value{
Returns a dataframe of the current gene list.
}
\description{
Load a PanSolid copy number variant gene table
}
