% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{make_cnv_triptych_plot}
\alias{make_cnv_triptych_plot}
\title{Create a triptych CNV plot with primer and exon tracks}
\usage{
make_cnv_triptych_plot(input_plot)
}
\arguments{
\item{input_plot}{The input plot to format as a triptych. This can be combined
with the make_labno_cnv_plot and make_fold_change_cnv_plot functions above
(see example).}
}
\value{
A triptych plot with 3 panels: first the input plot displaying the copy number
variant calls, then a plot showing the locations of PanSolid QIAseq primers, then
a plot showing the locations of target genes and their individual exons. All plots
have a consistently formatted X axis,
}
\description{
Create a triptych CNV plot with primer and exon tracks
}
