% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contamination_functions.R
\name{make_contamination_snp_plot}
\alias{make_contamination_snp_plot}
\title{Make a SNP plot to investigate contamination}
\usage{
make_contamination_snp_plot(
  s1_df,
  s1_id,
  s2_df,
  s2_id,
  hom_vaf_upper_threshold = 90,
  het_vaf_upper_threshold = 60,
  het_vaf_lower_threshold = 40,
  hom_vaf_lower_threshold = 10
)
}
\arguments{
\item{df}{The dataframe containing all the SNP data. This can be
created using the \code{read_snp_sheet} function.}

\item{sample1_ws}{The worksheet number of the sample in which
contamination is suspected.}

\item{sample1_labno}{The sample number (lab number) of the sample in which
contamination is suspected.}

\item{sample2_ws}{The worksheet number of the sample which is suspected to
be contaminating sample 1.}

\item{sample2_labno}{The sample number (lab number)  of the sample
which is suspected to be contaminating sample 1.}
}
\value{
A plot of the SNP comparisons.
}
\description{
Make a SNP plot to investigate contamination
}
