% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contamination_functions.R
\name{make_contamination_snp_plotlist}
\alias{make_contamination_snp_plotlist}
\title{Generate a plotlist of comparison SNP plots for a single sample}
\usage{
make_contamination_snp_plotlist(
  df,
  sample1_labno,
  sample1_ws,
  sample2_ws,
  plotlist_title
)
}
\arguments{
\item{df}{The dataframe containing all the SNP data.}

\item{sample1_labno}{The sample number (lab number) of the sample in which
contamination is suspected.}

\item{plotlist_title}{The title for the final plot file, which should end
in an appropriate suffix such as ".pdf"}

\item{ws}{The worksheet that \verb{sample1 labno} is on.}
}
\value{
This functions exports the plotlist as a multi-page file to your
local directory.
}
\description{
This function allows you to quickly investigate whether a sample shares a
similar SNP profile (and therefore is likely contaminated with) another
sample on the same PanSolid worksheet.
}
