% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_raw_excel_functions.R
\name{make_empty_cnv_df}
\alias{make_empty_cnv_df}
\title{Create an empty dataframe for CLC outputs with no CNVs}
\usage{
make_empty_cnv_df()
}
\value{
A dataframe with default columns for CLC exports
}
\description{
When no copy number variants (CNVs) are detected with a
sample, the tab of the raw CLC output is empty,
without column headings. This can make it difficult
when collating results from multiple files. This function
creates a placeholder dataframe with the correct
column headings to show that no CNVs have been identified.
}
