% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{make_exon_plot}
\alias{make_exon_plot}
\title{Make an plot showing exon data for inclusion in a CNV plot}
\usage{
make_exon_plot(plot_xmin, plot_xmax, interval, chromosome)
}
\arguments{
\item{plot_xmin}{The desired X axis minimum value}

\item{plot_xmax}{The desired X axis maximum value}

\item{interval}{The desired interval for X axis breaks}

\item{chromosome}{The chromosome for the gene of interest}
}
\value{
A plot showing the exon locations for target genes within the specified
genomic region.
}
\description{
Make an plot showing exon data for inclusion in a CNV plot
}
