% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{make_fold_change_cnv_plot}
\alias{make_fold_change_cnv_plot}
\title{Make a CNV plot with fold change on the Y axis}
\usage{
make_fold_change_cnv_plot(
  df,
  gene,
  interval = 10000,
  buffer = 5000,
  ymin = 0,
  ymax = 40,
  title = ""
)
}
\arguments{
\item{df}{The dataframe containing CNV information to be plotted.
This should be based on the "Positive CNV results" table from the PanSolid}

\item{gene}{The gene of interest}

\item{interval}{The desired interval for X axis breaks}

\item{buffer}{The desired buffer to add onto the width of the X axis.}

\item{ymin}{The desired Y axis minimum value}

\item{ymax}{The desired Y axis maximum value}
}
\value{
A list of values for the creation of other plots, and the rendered CNV plot.
These additional values are supplied so that the output of this function can act
an input to the make_cnv_triptych_plot function.
}
\description{
Make a CNV plot with fold change on the Y axis
}
