% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{make_labno_cnv_plot}
\alias{make_labno_cnv_plot}
\title{Make a CNV plot with sample lab number on the Y axis}
\usage{
make_labno_cnv_plot(
  df,
  gene,
  interval = 10000,
  buffer = 5000,
  yaxis = labno,
  title = "",
  yaxis_title = "Sample number"
)
}
\arguments{
\item{df}{The dataframe containing CNV information to be plotted.
This should be based on the "Positive CNV results" table from the PanSolid
Excel files.}

\item{gene}{The gene of interest}

\item{interval}{The desired interval for X axis breaks}

\item{buffer}{The desired buffer to add onto the width of the X axis.}

\item{yaxis}{The variable to plot on the Y axis, which defaults to labno, but can be
changed to other sample identifiers (such as NHS number) if desired.}
}
\value{
A list of values for the creation of other plots, and the rendered CNV plot.
These additional values are supplied so that the output of this function can act
an input to the make_cnv_triptych_plot function.
}
\description{
Make a CNV plot with sample lab number on the Y axis
}
