% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_functions.R
\name{make_primer_plot}
\alias{make_primer_plot}
\title{Make a plot showing PanSolid QIAseq primer location data for inclusion in a CNV plot}
\usage{
make_primer_plot(plot_xmin, plot_xmax, interval, chromosome)
}
\arguments{
\item{plot_xmin}{The desired X axis minimum value}

\item{plot_xmax}{The desired X axis maximum value}

\item{interval}{The desired interval for X axis breaks}

\item{chromosome}{The chromosome for the gene of interest}
}
\value{
A plot showing the locations of QIAseq primers within the specified
genomic region.
}
\description{
Make a plot showing PanSolid QIAseq primer location data for inclusion in a CNV plot
}
