% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_functions.R
\name{parse_filename}
\alias{parse_filename}
\title{Parse an element from a PanSolid results Excel filename}
\usage{
parse_filename(input_file, input_group)
}
\arguments{
\item{input_file}{The filename to parse}

\item{input_group}{The number of the group to extract, defined by filename_regex}
}
\value{
The string of the selected regex group
}
\description{
Parse an element from a PanSolid results Excel filename
}
\examples{
filename <- "Annotated_WS123456_12345678a_JoeShaw.xlsx"

worksheet <- parse_filename(filename, 1)
}
