% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{parse_wgs_cnv_class}
\alias{parse_wgs_cnv_class}
\title{Parse the CNV class from the WGS variant type string}
\usage{
parse_wgs_cnv_class(x)
}
\arguments{
\item{x}{The WGS variant type string from a WGS HTML which has the
variant type and then may also have the copy number in brackets.
Examples: "LOH(2)", "GAIN(3)", "INV"}
}
\value{
The CNV class (examples: "LOH", "GAIN", "INV") as a string
}
\description{
Parse the CNV class from the WGS variant type string
}
