% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{parse_wgs_cnv_copy_number}
\alias{parse_wgs_cnv_copy_number}
\title{Parse the CNV copy number from the WGS variant type string}
\usage{
parse_wgs_cnv_copy_number(x)
}
\arguments{
\item{x}{The WGS variant type string from a WGS HTML which has the
variant type and then may also have the copy number in brackets.
Examples: "LOH(2)", "GAIN(3)", "INV"}
}
\value{
The CNV copy number as a number
}
\description{
Parse the CNV copy number from the WGS variant type string
}
