% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{parse_wgs_html_grch38_coordinates}
\alias{parse_wgs_html_grch38_coordinates}
\title{Parse a group from the whole genome sequencing GRCh38 coordinate string}
\usage{
parse_wgs_html_grch38_coordinates(x, group)
}
\arguments{
\item{x}{The coordinate string. Example: "7:60912080-159335569"}

\item{group}{The group to be selected, which must be 1 for "chromosome",
2 for "first coordinate" or 4 for "second coordinate".}
}
\value{
The selected regex group as a string.
}
\description{
Parse a group from the whole genome sequencing GRCh38 coordinate string
}
