% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{parse_wgs_html_header}
\alias{parse_wgs_html_header}
\title{Parse the header element of a whole genome sequencing HTML}
\usage{
parse_wgs_html_header(html_filepath)
}
\arguments{
\item{html_filepath}{The full file-path of the HTML to parse}
}
\value{
The header identifiers as a data-frame
}
\description{
Parse the header element of a whole genome sequencing HTML
}
