% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{parse_wgs_html_pid_text}
\alias{parse_wgs_html_pid_text}
\title{Parse the patient identifier text from a whole genome sequencing HTML}
\usage{
parse_wgs_html_pid_text(html_filepath)
}
\arguments{
\item{html_filepath}{The full file-path of the HTML to parse}
}
\value{
The patient identifiers as a data-frame
}
\description{
Parse the patient identifier text from a whole genome sequencing HTML
}
\note{
This text appears as 3 rows above the referral ID table on a whole
genome sequencing HTML. The referral ID table itself needs to be parsed
using the parse_wgs_html_table_by_number function.
}
