% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{parse_wgs_html_table_by_div_id}
\alias{parse_wgs_html_table_by_div_id}
\title{Parse whole genome sequencing HTML files using a CSS "div_id" identifier.}
\usage{
parse_wgs_html_table_by_div_id(html_filepath, div_id)
}
\arguments{
\item{html_filepath}{The filepath for the WGS HTML file}

\item{div_id}{The CSS identifier for the table (examples below)}
}
\value{
Returns the table from the HTML as a tibble with column names
in snake-case.
}
\description{
Parse whole genome sequencing HTML files using a CSS "div_id" identifier.
}
\section{Useful Identifiers}{

"t_tumour_details", "t_tumour_sample",
"t_germline_sample", "t_quality", "tier1" (Tier 1 sequence variants),
"svcnv_tier1" (Tier 1 structural and copy number variants, for v2.28
and earlier),
"d_svcnv_tier1" (Tier 1 structural and copy number variants,later versions
than v2.28)
}

