% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{parse_wgs_html_table_by_number}
\alias{parse_wgs_html_table_by_number}
\title{Parse whole genome sequencing HTML files using the table number}
\usage{
parse_wgs_html_table_by_number(html_filepath, table_number)
}
\arguments{
\item{html_filepath}{The filepath for the WGS HTML file}

\item{table_number}{The number of the table to select}
}
\value{
Returns the table from the HTML as a tibble with column names
in snake-case.
}
\description{
Parse whole genome sequencing HTML files using the table number
}
\note{
This function is a less specific version of
parse_wgs_html_table_by_div_id.
It can be used for cases where the table does not have a div_id available.
This is specifically relevant to the patient details table - see example.
}
