% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erbb2_functions.R
\name{plot_timestamp}
\alias{plot_timestamp}
\title{Save a plot with a timestamp}
\usage{
plot_timestamp(
  input_plot,
  input_width = 15,
  input_height = 12,
  dpi = 300,
  folder
)
}
\arguments{
\item{input_plot}{Plot to save}

\item{input_width}{Desired plot width}

\item{input_height}{Desired plot height}

\item{dpi}{Desired plot resolution in dots per inch}

\item{folder}{The folder to save the plot into}
}
\value{
Saves the plot in the plots folder with a timestamp
}
\description{
Save a plot with a timestamp
}
\note{
This function is used for exporting plots for inclusions in validation
documentation. The default inputs are designed for presenting a plot as
half an A4 page.
}
