% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom_functions.R
\name{pool_sd}
\alias{pool_sd}
\title{Calculate the pooled standard deviation (pooled variance) for grouped standard deviation data}
\usage{
pool_sd(df)
}
\arguments{
\item{df}{A dataframe returned from the \link{group_sd} function.}
}
\value{
The pooled standard deviation value
}
\description{
Calculate the pooled standard deviation (pooled variance) for grouped standard deviation data
}
