% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{read_all_amp_genes_results}
\alias{read_all_amp_genes_results}
\title{Read the "All amplification genes" table from a PanSolid results Excel}
\usage{
read_all_amp_genes_results(file, sheet = "Amplifications")
}
\arguments{
\item{file}{The full filename of the Excel to read}

\item{sheet}{The sheet to read the table from}
}
\value{
The all amp gene CNV table as a dataframe with filename identifiers
}
\description{
Read the "All amplification genes" table from a PanSolid results Excel
}
