% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcr_functions.R
\name{read_biorad_ddpcr_csv}
\alias{read_biorad_ddpcr_csv}
\title{Read a CSV file from a BioRad ddPCR experiment as a data-frame}
\usage{
read_biorad_ddpcr_csv(filepath)
}
\arguments{
\item{filepath}{The full filepath of the CSV}
}
\value{
The data as a data-frame with column names in snakecase and
appropriate column types.
}
\description{
Read a CSV file from a BioRad ddPCR experiment as a data-frame
}
\note{
This function is designed to handle droplet digital polymerase chain reaction
(ddPCR) csv files exported from Quantasoft v1.7.4 (BioRad).
}
