% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{read_cnv_sheet}
\alias{read_cnv_sheet}
\title{Read the entire CNV sheet from a PanSolid CLC Excel file}
\usage{
read_cnv_sheet(filepath, sheet_regex = "CNVs_")
}
\arguments{
\item{filepath}{Full file path to an Excel file}

\item{sheet_regex}{Regular expression for sheet name matching}
}
\value{
The full CNV sheet as a tibble
}
\description{
The new PanSolid CLC Excel output includes multiple tables on the
same tab. \code{read_cnv_sheet} reads the entire sheet as one table, which
allows individual tables to be subsequently extracted
using \code{extract_cnv_tbls} without the file being read multiple times.
}
