% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{read_loh_table}
\alias{read_loh_table}
\title{Read the loss of heterozygosity table from PanSolid CNV Excels}
\usage{
read_loh_table(
  filepath,
  sheet_regex = "LOH_",
  sheet_range = "A1:G8",
  loh_genes = c("MSH2", "MSH6", "MLH1", "PMS2", "LZTR1", "SMARCB1", "NF2")
)
}
\arguments{
\item{filepath}{The full filepath to the Excel file}

\item{sheet_regex}{Regular expression for sheet name matching, defaults to
"LOH_".}

\item{loh_genes}{Vector of genes expected to have LOH results.}
}
\value{
The LOH results table as a tibble
}
\description{
Read the loss of heterozygosity table from PanSolid CNV Excels
}
