% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{read_percent_138_results}
\alias{read_percent_138_results}
\title{Read the percentage of the whole panel covered at 138X from a PanSolid Excel}
\usage{
read_percent_138_results(file, sheet = "Amplifications")
}
\arguments{
\item{file}{The full filename of the Excel to read}

\item{sheet}{The sheet to read from}
}
\value{
A dataframe of the filename identifiers with the percentage of
the whole panel covered at 138X value
}
\description{
Read the percentage of the whole panel covered at 138X from a PanSolid Excel
}
