% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{read_pos_cnv_results}
\alias{read_pos_cnv_results}
\title{Read the "Positive CNV results" table from a PanSolid results Excel}
\usage{
read_pos_cnv_results(file, sheet = "Amplifications")
}
\arguments{
\item{file}{The full filename of the Excel to read}

\item{sheet}{The sheet to read the table from}
}
\value{
The positive CNV table as a dataframe with filename identifiers
}
\description{
Read the "Positive CNV results" table from a PanSolid results Excel
}
