% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_arm_functions.R
\name{read_snp_sheet}
\alias{read_snp_sheet}
\title{Read the SNP frequency sheet from PanSolid Excels}
\usage{
read_snp_sheet(filepath, sheetname = "Artefacts_removed_GIAB_filt...")
}
\arguments{
\item{filepath}{The filepath of the Excel to read from.}

\item{sheetname}{The name of the sheet to read from.}
}
\value{
A dataframe of the first 12 columns of the Excel sheet, which is
annotated with the sample details from the filename.
}
\description{
This function reads the "Artefacts_removed_GIAB_filt..." sheet from the
Excel file which is saved in the "Unannotated_Files" folder of each
PanSolid worksheet. This sheet includes data for the single nucleotide
polymorphisms (SNPs) detected in the sample, which are presented in
the variant allele frequency (VAF) track of the interactive HTML file.
}
