% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pansolid_cnv_excel_functions.R
\name{read_stdev_results}
\alias{read_stdev_results}
\title{Read the standard deviation of signal-adjusted noise ratios from a PanSolid Excel}
\usage{
read_stdev_results(file, sheet = "Amplifications")
}
\arguments{
\item{file}{The full filename of the Excel to read}

\item{sheet}{The sheet to read from}
}
\value{
A dataframe of the filename identifiers with the standard deviation value
}
\description{
Read the standard deviation of signal-adjusted noise ratios from a PanSolid Excel
}
