% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{wgs_html_grch38_coordinates_regex}
\alias{wgs_html_grch38_coordinates_regex}
\title{Regular expression for the whole genome sequencing variant coordinates string}
\usage{
wgs_html_grch38_coordinates_regex()
}
\value{
The regular expression for the "Variant GRCh38 coordinates"
column in the WGS HTML file.
The format gives the chromosome, start coordinate and end coordinate for a CNV.
Example: "7:60912080-159335569"
}
\description{
Regular expression for the whole genome sequencing variant coordinates string
}
