% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_html_functions.R
\name{wgs_html_variant_type_regex}
\alias{wgs_html_variant_type_regex}
\title{Regular expression for whole genome sequencing "variant type" HTML column}
\usage{
wgs_html_variant_type_regex()
}
\value{
The regular expression for the "variant type" column
(example format: LOH(2); GAIN(3)).
This output is then used for later functions for extracting different
parts of the regex.
}
\description{
Regular expression for whole genome sequencing "variant type" HTML column
}
